Refractoring lownodeutilization - extracting common functions
